pyproject.toml
[tool.poetry]
name = "acs-token"
version = "0.1.0"
description = ""
authors = ["Eddie Lin <a-eddielin@microsoft.com>"]
readme = "README.md"
packages = [{include = "acs_token"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.104.1"}
azure-communication-identity = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
